Sat, Aug 19, 2023, 8:57 PM -05:00

+++ pset2/practice/hours/hours.c
@@ -29,7 +29,7 @@ float calc_hours(int hours[], int weeks, char output)
 {
     int total = 0;
 
-    for(int i = 0; i < weeks; i++)
+    for (int i = 0; i < weeks; i++)
     {
         total += hours[i];
     }
@@ -40,6 +40,6 @@ float calc_hours(int hours[], int weeks, char output)
     }
     else
     {
-        return total / weeks;
+        return (float) (total / weeks);
     }
 }
